pyproject.toml
[tool.poetry]
name = "trashbot"
version = "0.1.0"
description = "An AI Assistant in the form of a trashbot."
authors = ["Reuben Lewis <reuben.lewis@outlook.com>"]
license = "MIT"

[tool.poetry.scripts]
parsing = 'trashbot.cli.commands:test_parsing'
create-table = 'trashbot.cli.commands:create_reminder_table'

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
click = "^8.1.3"
adapt-parser = {git = "https://github.com/mycroftai/adapt.git"}
Cython = "^0.29.30"
scipy = "^1.8.1"
h5py = "^3.7.0"
asyncio = "^3.4.3"
flake8 = "^4.0.1"
black = "^22.6.0"
PyYAML = "^6.0"
attrs = "^21.4.0"
dateparser = "^1.1.1"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
